feat : 근접 검 레벨업
근접 검 레벨업 선택 시 데미지와 갯수가 늘어남. 캐릭터 근처 풀링으로 비활성화 시켜둔 후 레벨업 시 활성화 되도록 구현.
아이템 레벨업 함수는 데이터매니저에서 관리

+++ Assets/Scripts/Items/ItemData.cs
@@ -23,7 +23,7 @@ public class ItemData : ScriptableObject
         public float baseDamage;        // 기본 데미지
         public float[] damages;
         public int baseCount;           // 기본 Count (갯수)
-        public float[] counts;
+        public int[] counts;
 
         public int currentLevel;        // 아이템 현재 레벨
         public int maxLevel;            // 아이템 최대 레벨
@@ -32,7 +32,7 @@ public class ItemData : ScriptableObject
         public float cooldown;          // 쿨타임
         public float duration;          // 공격 지속 시간 (발사체 유지 시간)
 
-        public float count;             // 갯수
+        public int count;             // 갯수
         public float size;              // 크기
         public float speed;             // 속도
         public float pierce;            // 관통 횟수
